rename memory card game folder
