+++ client/src/app/app-routing.module.ts
@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { GamePageComponent } from './game-page/game-page.component';
+import { GamePageComponent } from './match-cards-game/game-page.component';
 import { HomePageComponent } from './home-page/home-page.component';
 
 const routes: Routes = [
   {path: '', component: HomePageComponent, pathMatch: 'full'},
-  {path: 'game', component: GamePageComponent},
+  {path: 'match-cards-game', component: GamePageComponent},
   {path: '**', redirectTo: ''},
 ];
