+++ client/src/app/app-store.model.ts
@@ -1,5 +1,5 @@
 import { AuthUserState } from "./app-shell/store/reducer/auth.reducer";
-import { GameResultState } from "./game-page/store/reducers/game-result.reducer";
+import { GameResultState } from "./match-cards-game/store/reducers/game-result.reducer";
 
 export interface AppStore {
     authUser: AuthUserState;
