+++ client/src/app/home-page/home-page.component.scss
@@ -3,12 +3,13 @@
     justify-content: center;
     align-items: center;
     height: 100%;
-}
-
-.game-card {    
-
+  }
+  
+  .game-card {    
+    margin: 1rem;
+    
     mat-card-header {
-        justify-content: center;
+      justify-content: center;
     }
 
     .card-image {
