+++ client/src/app/match-cards-game/config/dino.config.ts
