+++ client/src/app/match-cards-game/config/game-settings.config.ts
