+++ client/src/app/match-cards-game/game-page.component.scss
