+++ client/src/app/match-cards-game/models/card-image-data.model.ts
