+++ client/src/app/match-cards-game/models/memory-game-result.model.ts
