+++ client/src/app/match-cards-game/services/card-config-generator.service.ts
