+++ client/src/app/match-cards-game/services/game-result.service.ts
