+++ client/src/app/match-cards-game/store/effects/game-result.effect.ts
